data/requirements.txt
et-xmlfile==1.1.0
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.2
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
